Move everything around, fix debug issue
